Consolidate ROS 1 & ROS 2 package manifests
Previously, this package had separate package manifest files for ROS 1 and ROS 2, and it was necessary to run a script after cloning the repository to name the correct one package.xml.

This consolidates both of them into a single package.xml file that works with both ROS 1 and ROS 2, so the package can now be cloned and built inside a standard colcon workspace without needing to run any additional scripts.  It also makes a few tweaks to other files to ensure compatibility.

This has been tested on ROS Noetic and ROS Foxy in Ubuntu 20.04.