Fix ROS 1 build
catkin_make was refusing to build this package because it couldn't
identify the build type, so this explicitly sets the build_type to
catkin when in a ROS 1 environment.

Also, there was a header that was defining some values that should
only be set in ROS 2, and the #if definition around it was accidentally
checking if the ROS version was >0 rather than >1, so this also fixes
that.

Signed-off-by: P. J. Reed <phillipreed@hatchbed.com>